requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
cssselect==1.2.0
decorator==5.1.1
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.7
django-crispy-forms==2.1
django-email==0.1.10
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.0
gTTS==2.5.0
idna==3.4
jieba3k==0.35.1
joblib==1.3.2
lxml==5.1.0
Markdown==3.6
MouseInfo==0.1.3
mss==9.0.1
mysql-connector-python==8.2.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.0
opencv-contrib-python==4.8.0.76
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==2.2.0
Pillow==10.0.1
platformdirs==3.11.0
praw==7.7.1
prawcore==2.4.0
protobuf==4.21.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyAutoGUI==0.9.54
pyenchant==3.2.2
pygame==2.5.2
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
ratelim==0.1.6
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
scipy==1.11.3
sgmllib3k==1.0.0
six==1.16.0
sketchpy==0.1.7
soupsieve==2.5
sqlparse==0.4.4
svg.path==6.3
svgpathtools==1.6.1
svgwrite==1.4.3
tinysegmenter==0.3
tldextract==5.1.1
tqdm==4.66.1
turtle==0.0.1
typing_extensions==4.11.0
tzdata==2023.3
update-checker==0.18.0
urllib3==2.0.5
virtualenv==20.24.6
websocket-client==1.7.0
